+++ src/rex.forms/static/js/lib/elements/readOnlyWidgetTypes.js
@@ -14,6 +14,7 @@ var readOnlyWidgetTypes = {
 
   'text': [
     'readOnlyInputText',
+    'readOnlyLookupText',
     'readOnlyTextArea'
   ],