+++ src/rex.forms/static/js/lib/elements/widgetTypes.js
@@ -14,6 +14,7 @@ var widgetTypes = {
 
   'text': [
     'inputText',
+    'lookupText',
     'textArea'
   ],